Move shielded_pool and stake into component crate
Cargo fmt

Clippy lints
