+++ component/Cargo.toml
@@ -6,10 +6,39 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-tokio = { version = "1.16", features = ["full", "tracing"]}
+# Workspace dependencies
+penumbra-crypto = { path = "../crypto" }
+penumbra-proto = { path = "../proto" }
 penumbra-transaction = { path = "../transaction" }
+penumbra-storage = { path = "../storage" }
 penumbra-chain = { path = "../chain" }
+penumbra-tct = { path = "../tct" }
+
+# Penumbra dependencies
+tendermint = "0.24.0-pre.1"
+decaf377 = { git = "https://github.com/penumbra-zone/decaf377" }
+jmt = { git = "https://github.com/penumbra-zone/jellyfish-merkle.git", branch = "main" }
+
+tokio = { version = "1.16", features = ["full", "tracing"]}
 async-trait = "0.1.52"
 tonic = "0.6.1"
-tendermint = "0.24.0-pre.1"
 anyhow = "1"
+tracing = "0.1"
+ark-ff = "0.3"
+blake2b_simd = "0.5"
+bincode = "1.3.3"
+serde = { version = "1", features = ["derive"] }
+metrics = "0.18"
+sha2 = "0.9"
+serde_json = "1"
+bech32 = "0.8"
+regex = "1.5"
+once_cell = "1.8"
+bitvec = "1"
+
+[dev-dependencies]
+ed25519-consensus = "2"
+rand_core = "0.6"
+
+[build-dependencies]
+vergen = "5"
