+++ component/src/shielded_pool/metrics.rs
