+++ component/src/shielded_pool/mod.rs
@@ -1,12 +1,9 @@
-// Required to ensure that Rust can infer a Send bound inside the TCT
-#![recursion_limit = "256"]
-
 mod commission;
 mod component;
 pub(crate) mod event;
 mod metrics;
 pub(crate) mod state_key;
 
-pub use crate::metrics::register_metrics;
+pub use self::metrics::register_metrics;
 pub use commission::{CommissionAmount, CommissionAmounts};
 pub use component::{ShieldedPool, View};
