+++ component/src/stake/validator/definition.rs
@@ -2,7 +2,7 @@ use penumbra_crypto::rdsa::{Signature, SpendAuth};
 use penumbra_proto::{stake as pb, Protobuf};
 use serde::{Deserialize, Serialize};
 
-use crate::validator::Validator;
+use crate::stake::validator::Validator;
 
 /// Authenticated configuration data for a validator.
 #[derive(Debug, PartialEq, Eq, Clone, Serialize, Deserialize)]
