+++ custody/Cargo.toml
@@ -10,7 +10,6 @@ edition = "2021"
 penumbra-proto = { path = "../proto" }
 penumbra-chain = { path = "../chain" }
 penumbra-crypto = { path = "../crypto" }
-penumbra-stake = { path = "../stake" }
 penumbra-transaction = { path = "../transaction" }
 
 tokio = { version = "1.16", features = ["full"]}
