+++ pd/src/components/app.rs
@@ -2,10 +2,10 @@ use anyhow::Result;
 use async_trait::async_trait;
 use jmt::{RootHash, Version};
 use penumbra_chain::{genesis, View as _};
+use penumbra_component::shielded_pool::ShieldedPool;
+use penumbra_component::stake::component::Staking;
 use penumbra_component::{Component, Context};
 use penumbra_ibc::IBCComponent;
-use penumbra_shielded_pool::ShieldedPool;
-use penumbra_stake::component::Staking;
 use penumbra_storage::{State, StateExt, Storage};
 use penumbra_transaction::Transaction;
 use tendermint::abci::{self, types::ValidatorUpdate};
