+++ pd/src/metrics.rs
@@ -19,9 +19,9 @@ pub use metrics::*;
 /// functions in our dependencies.
 pub fn register_metrics() {
     penumbra_storage::register_metrics();
-    penumbra_stake::register_metrics();
+    penumbra_component::stake::register_metrics();
     penumbra_ibc::register_metrics();
-    penumbra_shielded_pool::register_metrics();
+    penumbra_component::shielded_pool::register_metrics();
 
     register_counter!(MEMPOOL_CHECKTX_TOTAL);
     describe_counter!(
