--- stake/Cargo.toml
@@ -1,44 +1,0 @@
-[package]
-name = "penumbra-stake"
-version = "0.1.0"
-authors = ["Penumbra Labs <team@penumbra.zone>"]
-edition = "2021"
-description = "The staking and delegation implementation for Penumbra"
-repository = "https://github.com/penumbra-zone/penumbra/"
-homepage = "https://penumbra.zone"
-license = "MIT OR Apache-2.0"
-publish = false
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-# Workspace dependencies
-penumbra-component = { path = "../component" }
-penumbra-crypto = { path = "../crypto" }
-penumbra-proto = { path = "../proto" }
-penumbra-transaction = { path = "../transaction" }
-penumbra-shielded-pool = { path = "../shielded-pool" }
-penumbra-chain = { path = "../chain" }
-penumbra-storage = { path = "../storage" }
-
-# Penumbra dependencies
-tendermint = "0.24.0-pre.1"
-# External dependencies
-anyhow = "1"
-serde_json = "1"
-serde = { version = "1", features = ["derive"] }
-bech32 = "0.8"
-regex = "1.5"
-once_cell = "1.8"
-bitvec = "1"
-async-trait = "0.1.52"
-sha2 = "0.9"
-tracing = "0.1"
-metrics = "0.18"
-
-[dev-dependencies]
-ed25519-consensus = "2"
-rand_core = "0.6"
-
-[build-dependencies]
-vergen = "5"
