+++ wallet/Cargo.toml
@@ -16,8 +16,8 @@ publish = false
 penumbra-proto = { path = "../proto" }
 penumbra-chain = { path = "../chain" }
 penumbra-crypto = { path = "../crypto" }
-penumbra-stake = { path = "../stake" }
 penumbra-transaction = { path = "../transaction" }
+penumbra-component = { path = "../component" }
 penumbra-tct = { path = "../tct" }
 penumbra-view = { path = "../view" }
 penumbra-custody = { path = "../custody" }
