+++ wallet/src/plan.rs
@@ -1,13 +1,13 @@
 use std::collections::HashMap;
 
 use anyhow::Result;
+use penumbra_component::stake::rate::RateData;
+use penumbra_component::stake::validator;
 use penumbra_crypto::{
     asset::Denom, keys::DiversifierIndex, memo::MemoPlaintext, Address, DelegationToken,
     FullViewingKey, Value, STAKING_TOKEN_ASSET_ID, STAKING_TOKEN_DENOM,
 };
 use penumbra_proto::view::NotesRequest;
-use penumbra_stake::rate::RateData;
-use penumbra_stake::validator;
 use penumbra_transaction::{
     plan::{ActionPlan, OutputPlan, SpendPlan, TransactionPlan},
     Fee,
